게시물 수정시 지우는 파일 없는 경우 NullPointerException 발생 -> 해결

+++ backend/src/main/java/com/example/backend/service/borad/BoardService.java
@@ -133,18 +133,20 @@ public class BoardService {
     }
 
     public boolean update(Board board, List<String> removeFiles) {
-        for (String file : removeFiles) {
-            String key = STR."prj1114/\{board.getId()}/\{file}";
-            DeleteObjectRequest dor = DeleteObjectRequest.builder()
-                    .bucket(bucketName)
-                    .key(key)
-                    .build();
+        if (removeFiles != null) {
+            for (String file : removeFiles) {
+                String key = STR."prj1114/\{board.getId()}/\{file}";
+                DeleteObjectRequest dor = DeleteObjectRequest.builder()
+                        .bucket(bucketName)
+                        .key(key)
+                        .build();
 
-            // s3 파일 지우기
-            s3.deleteObject(dor);
+                // s3 파일 지우기
+                s3.deleteObject(dor);
 
-            // db 파일 지우기
-            mapper.deleteFileByBoardIdAndName(board.getId(), file);
+                // db 파일 지우기
+                mapper.deleteFileByBoardIdAndName(board.getId(), file);
+            }
         }
